requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
coloredlogs==14.0
discord.py==1.4.1
ffmpeg==1.4
humanfriendly==8.2
idna==2.10
multidict==4.7.6
pycparser==2.20
PyNaCl==1.3.0
pyreadline==2.1
python-dotenv==0.14.0
six==1.15.0
yarl==1.5.1
youtube-dl==2020.9.20
